Fix over-capitalization of <h3> headings
It is generally considered good style not to necessarily capitalize
every single word in titles, e.g.

  https://grammar.yourdictionary.com/capitalization/rules-for-capitalization-in-titles.html

Presumably this is intended to include sub-headings too.

In fact, the OpenStack Documentation Contributor Guide goes even
further away from all capitals:

    Use the following guidelines for all types of headings:

      * Use sentence-style capitalization.
      * ...

    -- https://docs.openstack.org/doc-contrib-guide/writing-style/headings.html

And looking at files such as www/stein/user/index.html, we find headings
such as:

    User Guides for OpenStack Services

and

    User Guides for OpenStack Deployment Tools

where the intention for a lowercase "for" is clear.

So remove the CSS which automatically capitalizes every single word in
headings, and lowercase any remaining articles, prepositions, or
coordinating conjunctions.

Change-Id: I7cc35a1528c0e15c4e1d1fc47d27c2d4ab49e497