Provide dropdown menu which allows switching doc release
Currently deprecation badge headers just point back to the root of
the documentation for the latest stable release, so for example the
badge at the top of

    https://docs.openstack.org/latest/user/index.html

just points to

    https://docs.openstack.org/rocky/

It's useful to be able to quickly jump between different releases of
the same documentation page, but currently this requires hand-editing
URLs or many extra manual hops in the navigation trail.  So
automatically generate the links, but only display them if the link
target exists.

Since the main portal page for each release already had a "More
Releases" dropdown, remove that to avoid two dropdowns on the portal
pages.

Change-Id: I2498f00a6c863d078a70289a655b0aa3958325ed