Fixing the "where to next?" buttons
The buttons underneath "What Are The Next Steps?" work for a
brief moment (clickable and workable) but then they
are 'greyed' out and you are unable to click. This is because
the use of float was taking the buttons out of the normal
flow of the page.

This fix places the buttons inside a div with the
"clearfix" class[1].

[1] https://getbootstrap.com/docs/4.0/utilities/clearfix/

Change-Id: I31d9d8a1e064a7a930de45682b3eaf1a35ece525
Closes-bug: 1811964