Fix per-project deprecation badges
I2498f00a6c863d078a70289a655b0aa3958325ed changed the badge shown at
the top of documentation pages which describes the state of the
release of the documentation currently being viewed.  However the
unfortunately release-switcher dropdown it introduced into
deprecated_badge.tmpl had the side-effect of affecting the badge.html
which is dynamically included in project documentation by some
Javascript in theme/openstackdocs/layout.html within the
openstackdocstheme repository, and the new release switcher was only
designed to work with openstack-manuals, not with any other projects.

This broke the badge for all releases (www/$RELEASE/badge.html),
excluding www/latest/ which has a hard-coded badge.html.  This was
missed during testing because the dynamic inclusion mechanism is
hardcoded to retrieve the badge from https://docs.openstack.org rather
than any local repository.

For a quick fix, reintroduce the old badge template as

    www/templates/project_deprecated_badge.tmpl

and change all the per-project badges to use this.  Later we can
investigate the possibility of enhancing this per-project badge so
that it provides release-switching functionality like the badge
currently in openstack-manuals.

Change-Id: I53c4b35e31bcdda16ffd50f9a8a26c773c8d3574