Use white'ish text for "eol", "obsolete" deprecated badges
The deprecated badge (which should really be called version banner or
something, given that it's not really a "badge" and appears for
non-deprecated releases too) is a simple div with some text, a close
button, and a Bootstrap dropdowns [1] to switch between versions.
This div has three classes, 'deprecated-badge', 'fixed', and
'deprecated-badge-X', where 'X' is the series status (i.e.
'deprecated-badge-current').

We were setting the background color of the div using the
'.deprecated-badge' selector and then overriding this using the
version-specific class selector. This background color was inherited by
the dropdown.

This had the end result of giving a unique color to the overall div for
different releases but the dropdown, which wasn't being overridden on a
per-version basis, stayed the same, yielding a consistent blue dropdown
regardless of the background.

We fix this by setting the background-color attributes on the dropdown
buttons explicitly, removing the background-color from the
'deprecated-badge' class. While we're at it, we also fix the text color
for the main div, so that we're not trying to read dark blue text on a
dark red background.

[1] https://getbootstrap.com/docs/4.0/components/dropdowns/

Change-Id: I8dbf54ff90164b2c2945ba74fee4bcf51eaecabb
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>