Use opendev-tox-docs job
Switch our docs job to use opendev-tox-docs job as base so that we can
publish later the content from gate directly (via promote queue) and do
not need to rebuild in post again.

Use the new promote-openstack-manuals job as first step in using promote
publishing. The other publish jobs will be replaced separately.

Use for building now publishdocs/publishlang so that we have exactly
the same content in check and gate - and in the published website.
For this, update tox.ini and publish script.

Also, do not publish project-data.

Depends-On: https://review.opendev.org/676489
Change-Id: I91f22173a4a6ed421f339833c486a31fd319dba3