Migrate from install to deploy-guide
The deployment guide is currently pointed at triplo-docs but it has been
requested that we actually publish a deployment guide. Efforts to
extracts many of the installation doc pages and moves them into the
deploy-guide source tree(/project-deploy-guide/tripleo-docs/latest/) is
already done. The deploy-guide was earlier published via
https://review.opendev.org/#/c/677029/.

This commit will migrate references to correctly point to the new deploy-guide.

Change-Id: I4378087fa0bc70dbeb1e948815c2eb0dbdd2fe4c