Update openstackdocstheme version
Switch to openstackdocstheme 2.2.0 version.

Update Sphinx version as well.

openstackdocstheme renames some variables, so follow the renames. A
couple of variables are also not needed anymore, remove them.

Set openstackdocs_auto_version to not version the documents and
openstackdocs_auto_name to use project as name.

Set openstackdocs_pdf_link to link to PDF files for the two
guides that build PDFs. Note that
the link to the published document only works on docs.openstack.org
where the PDF file is placed in the top-level html directory. The
site-preview places the PDF in a pdf directory.

Update copyright.

Change-Id: Ibd4844364d78abd135e5b094beba8f17503fab26