Exclude older ironic releases from indexing
A frequent complaint of those searching for information is they get
keyword hits on much older documentation indexes. Until very recently,
this often tended to be a more recent release, but sometiems liberty,
newton, or even ocata would appear in the results.

So this is an attempt to restrict indexing down to newer more relevant
versions to ultimately limit confusion while still keeping older docs
available for reference if needed.

Change-Id: I680a4a9165178b19b4d4b715d164f0e6f4f2e1d2