[www] Add Xena index pages
This change prepares the Xena release by adding the index pages. It
still keeps Wallaby as the current release. We also add the files needed
for the next release, Yoga. These are not published since we aren't
adding the corresponding index page yet.

A future change will transition Xena to maintained.

Change-Id: Ied5cb4a103569fa7159d8556c6baadac0c0b5a09
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>