Drop explicit latest from openinfra redirects
The redirects for refstack and python-tempestconf, which moved into
the openinfra namespace on docs.opendev.org, does not need to
specify latest on the target as the old and new paths both used
similar redirects already (unlike docs from the infra subtree, which
was likely the source of the confusion here). Drop the use of latest
for these, since otherwise old bookbarks or search results could
result in doubling the latest when the redirect on the new side is
applied.

Change-Id: Ieffb7ce6d84888588ee3c51388923b6862136bca