Add Cinder NVMe-oF target port
Similar to the iSCSI port that requires a firewall rule when using
Cinder LVM driver with an iSCSI target we also require port 4420 to be
allowed when using LVM with the NVMe-oF target.

This patch adds the 4420 port to the list of required ports and
clarifies a bit the iSCSI one, since it's only required when using LVM
with iSCSI, not always.

Change-Id: I499f1916eadc0f99558e529be6cc49576224c8f5