Fix 2023.1 Antelope sort order
We have wrapped the alphabet (and added a numeric prefix) which means
new releases now sort as older than old releases like Zed. This causes
the docs to report the latest release as maintained but older than Zed
which it believes is the most recent release.

We fix this be not calling sorted() on our dict entries and instead
expect us to maintain an explicit sort order using an OrderedDict. This
requires us to maintain the order but it appears we were doing this
already so no big deal.

Change-Id: I003f92a30fed5561dece60552929961cf910907d