Attempt at fixing broken docs
This patch attempts to fix the broken docs by having the series
name match the branch name (2023.1 instead of 2023.1.antelope)

A patch was previously merged and reverted[0]. It changed the
name of the series, but missed updating the directory names
under the www directory.

I did test this locally, and it did correctly pupulate the pages
with links for the various projects, but I do not know if it
fully resolves the issue.

Additionally adds redirects for /2023.1.antelope to /2023.1
and 2023.2.bobcat to 2023.2.

[0]. https://review.opendev.org/c/openstack/openstack-manuals/+/880060

Change-Id: I39e0740e57502fd96299fea79409e55e49c10367