[www] Set 2023.1 Antelope state to Unmaintained
2023.1 Antelope series reached it's end of maintained state with the
planned date of October 30th, 2024.

As most of the projects are ready with the transition [1], this patch
sets 2023.1 Antelope as unmaintained.

Note that stable/2023.1 branches are replaced with unmaintained/2023.1
branches and they are still open for bug fix backports, but NOT
supported or maintained by the project teams themselves (for details
about the Unmaintained state, see: [2][3]).

[1] https://review.opendev.org/q/topic:antelope-unmaintained
[2] https://governance.openstack.org/tc/resolutions/20230724-unmaintained-branches.html
[3] https://docs.openstack.org/project-team-guide/stable-branches.html

Change-Id: Ib1204f14682b59f17c0bd737de45db9a34dd1731