[www] Update project data of 2024.2 Dalmatian
Project data for a release is added just before a series final release.
At that time not all project merges all the necessary patches, hence
documentation is not available for all projects. As time passes some of
the originally missing documentation are already generated, so the
project data can be updated to the given release.

Change-Id: I2803b7b4ce5673f41c87b7dcb41c95b45e1777e6