[www] Update project data of 2025.1 Epoxy
Project data for a release is added just before a series final release.
At that time not all project merges all the necessary patches, hence
documentation is not available for all projects. As time passes some of
the originally missing documentation are already generated, so the
project data can be updated to the given release.

Change-Id: I086096e0ace110da2f8a4a5d273a370a7b9e55a9
Signed-off-by: Elod Illes <elod.illes@est.tech>