[www] Set 2024.1 Caracal state to Unmaintained
2024.1 Caracal series reached it's end of maintained state with the
planned date of October 24th, 2025.

As every (non-cycle-trailing) projects are ready with the transition to
Unmaintained [1], this patch updates 2024.1 Caracal state on docs [2]
site.

Note that stable/2024.1 branches are replaced with unmaintained/2024.1
branches and they are still open for bug fix backports, but NOT
supported or maintained by the project teams themselves (for details
about the Unmaintained state, see: [3][4]).

[1] https://review.opendev.org/q/topic:caracal-unmaintained
[2] https://docs.openstack.org
[3] https://governance.openstack.org/tc/resolutions/20230724-unmaintained-branches.html
[4] https://docs.openstack.org/project-team-guide/stable-branches.html

Change-Id: I4f582c80a2d0d08bf8c33b1d136a0362911dfcae
Signed-off-by: Elod Illes <elod.illes@est.tech>