Re-direct tripleo docs to repos
We retired tripleo in 2023 [1]. The team
still maintains stable branches where appropriate
but the "latest" links can be redirected to
their respective git repos which contain a
notice of retirement.

Alongside, also fix the template to generate
release specific docs to not include tripleo.

Doing this will help us not host
https://docs.openstack.org/os-net-config/latest
since the project is being handed over to external
maintainers [2]

[1] https://review.opendev.org/c/openstack/governance/+/905145
[2] https://review.opendev.org/c/openstack/governance/+/967463
Change-Id: I14d710bb002b4919374b3bedbbc505cf1c394dd2
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>